chore: Migrate CJS to ESM (#42072)
* chore: Migrate CJS to ESM

* fix

* fix

* fix

* fix

* fix

* update
