+++ package.json
@@ -53,7 +53,7 @@
     "check-ts-demo": "ts-node --esm --project tsconfig.node.json scripts/check-ts-demo.ts",
     "clean": "antd-tools run clean && rm -rf es lib coverage dist report.html",
     "clean-lockfiles": "rm -rf package-lock.json yarn.lock",
-    "collect-token-statistic": "ts-node --project tsconfig.node.json scripts/collect-token-statistic.js",
+    "collect-token-statistic": "ts-node --esm --project tsconfig.node.json scripts/collect-token-statistic.ts",
     "token-meta": "node scripts/generate-token-meta.js",
     "prestart": "npm run version && npm run collect-token-statistic && npm run token-meta",
     "precompile": "npm run prestart",
@@ -75,7 +75,7 @@
     "lint:demo": "eslint components/*/demo/*.md",
     "lint:deps": "antd-tools run deps-lint",
     "lint:md": "remark . -f -q",
-    "lint:style": "ts-node --project tsconfig.node.json scripts/check-cssinjs",
+    "lint:style": "ts-node --esm --project tsconfig.node.json scripts/check-cssinjs.ts",
     "lint:script": "eslint . --ext .js,.jsx,.ts,.tsx --cache",
     "pre-publish": "npm run test-all -- --skip-build && node ./scripts/pre-publish-notice.js",
     "prettier": "prettier -c --write **/* --cache",
