+++ scripts/check-cssinjs.ts
@@ -1,20 +1,20 @@
-/* eslint-disable import/no-unresolved,no-console,global-require,import/no-dynamic-require */
-const chalk = require('chalk');
-const {
+/* eslint-disable no-console */
+import {
   StyleProvider,
-  logicalPropertiesLinter,
   legacyNotSelectorLinter,
+  logicalPropertiesLinter,
   parentSelectorLinter,
-} = require('@ant-design/cssinjs');
-const React = require('react');
-const ReactDOMServer = require('react-dom/server');
-const { generateCssinjs } = require('./generate-cssinjs');
+} from '@ant-design/cssinjs';
+import chalk from 'chalk';
+import React from 'react';
+import ReactDOMServer from 'react-dom/server';
+import { generateCssinjs } from './generate-cssinjs';
 
 console.log(chalk.green(`🔥 Checking CSS-in-JS...`));
 
 let errorCount = 0;
 const originError = console.error;
-console.error = (msg) => {
+console.error = (msg: any) => {
   if (msg.includes('Warning: [Ant Design CSS-in-JS]')) {
     errorCount += 1;
     console.log(chalk.red(`❌ `), msg.slice(msg.indexOf('Error in')).replace(/\s+/g, ' '));
@@ -25,7 +25,7 @@ console.error = (msg) => {
 
 generateCssinjs({
   key: 'check',
-  render: (Component) => {
+  render(Component: any) {
     ReactDOMServer.renderToString(
       React.createElement(
         StyleProvider,
